[FIX] Fixed OpenMP error on Windows in render-based-tutorial-utils.h
On Windows, OpenMP does not support uint

+++ tutorial/tracking/render-based/render-based-tutorial-utils.h
@@ -207,7 +207,7 @@ public:
 #ifdef VISP_HAVE_OPENMP
 #pragma omp parallel for
 #endif
-      for (unsigned int i = 0; i < IRGB.getHeight(); ++i) {
+      for (int i = 0; i < static_cast<int>(IRGB.getHeight()); ++i) {
         memcpy(Iout[i], IgrayOverlay[i], IRGB.getWidth() * sizeof(vpRGBa));
         memcpy(Iout[i] + IRGB.getWidth(), IColOverlay[i], IRGB.getWidth() * sizeof(vpRGBa));
         memcpy(Iout[i + IRGB.getHeight()], IdepthOverlay[i], IRGB.getWidth() * sizeof(vpRGBa));
